client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.55.0",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "google-generativeai>=0.8.0",
]
